+++ server/routes.js
@@ -1,5 +1,5 @@
 // In-house dependencies
-const { clickUpHelpers } = require('brave-alert-lib')
+const { googleHelpers } = require('brave-alert-lib')
 const dashboard = require('./dashboard')
 const pa = require('./pa')
 const rak = require('./rak')
@@ -16,8 +16,8 @@ function configureRoutes(app) {
   app.get('/vitals', dashboard.sessionChecker, dashboard.renderVitalsPage)
 
   app.post('/login', dashboard.submitLogin)
-  app.post('/pa/aws-device-registration', pa.validateAwsDeviceRegistration, clickUpHelpers.clickUpChecker, pa.handleAwsDeviceRegistration)
-  app.post('/pa/buttons-twilio-number', pa.validateButtonsTwilioNumber, clickUpHelpers.clickUpChecker, pa.handleButtonsTwilioNumber)
+  app.post('/pa/aws-device-registration', pa.validateAwsDeviceRegistration, googleHelpers.paAuthorize, pa.handleAwsDeviceRegistration)
+  app.post('/pa/buttons-twilio-number', pa.validateButtonsTwilioNumber, googleHelpers.paAuthorize, pa.handleButtonsTwilioNumber)
   app.post('/rak_button_press', rak.validateButtonPress, rak.handleButtonPress)
 
   app.post('/api/message-clients', api.validateMessageClients, api.authorize, api.messageClients)